+++ test/DataView.test.js
@@ -28,6 +28,7 @@ describe('DataView', function () {
       {id: 2, content: 'Item 2', group: 2},
       {id: 3, content: 'Item 3', group: 2}
     ]);
+    assert.equal(group2.length, 2);
 
 // test filtering the view contents
     assert.deepEqual(group2.get({
@@ -51,19 +52,46 @@ describe('DataView', function () {
     groups.update({id:2, content: 'Item 2 (changed)'});
     assert.equal(groupsTriggerCount, 1);
     assert.equal(group2TriggerCount, 1);
+    assert.equal(group2.length, 2);
 
     groups.update({id:5, content: 'Item 5 (changed)'});
     assert.equal(groupsTriggerCount, 2);
     assert.equal(group2TriggerCount, 1);
+    assert.equal(group2.length, 2);
 
 // detach the view from groups
     group2.setData(null);
     assert.equal(groupsTriggerCount, 2);
     assert.equal(group2TriggerCount, 2);
+    assert.equal(group2.length, 0);
 
     groups.update({id:2, content: 'Item 2 (changed again)'});
     assert.equal(groupsTriggerCount, 3);
     assert.equal(group2TriggerCount, 2);
+
+    // test updating of .length property
+    group2.setData(groups);
+    assert.equal(group2.length, 2);
+
+    // add a new item
+    groups.add({id: 6, content: 'Item 6', group: 2});
+    assert.equal(group2.length, 3);
+
+    // change an items group to 2
+    groups.update({id: 4, group: 2});
+    assert.equal(group2.length, 4);
+
+    // change an items group to 1
+    groups.update({id: 4, group: 1});
+    assert.equal(group2.length, 3);
+
+    // remove an item
+    groups.remove(2);
+    assert.equal(group2.length, 2);
+
+    // remove all items
+    groups.clear();
+    assert.equal(group2.length, 0);
   });
 
 });