+++ HISTORY.md
@@ -6,7 +6,8 @@ http://visjs.org
 
 ### DataSet
 
-- Added property `length` holding the total number of items in the DataSet.
+- Added property `length` holding the total number of items to the `DataSet`
+  and `DataView`.
 
 
 ## 2015-01-16, version 3.9.1